installation_proxy: Ignore non-status messages instead of terminating loop
Some iOS versions will interleave a CFBundleIdentifier message into the Status
messages, and the code would treat the CFBundleIdentifier message as an error.
Terminating the loop, never seeing the last message - and even worse - never
calling back to ideviceinstaller that would be stuck waiting for a callback
with a message where Status == Complete.