socket: Fix IPv6 scope id lookup logic to handle another network device problem
The lookup logic preferred to return the last suitable scope id match.
This became a problem if there was already a suitable scope id match before
that was higher in the interface list. This now chooses the higher last
scope id interface match and thus probably in the routing preference.